# Stop parsing column XPath for each row