before

```
[info] Running (fork) org.openjdk.jmh.Main -i 3 -wi 3 -f1 -t 1
[info] # JMH version: 1.19
[info] # VM version: JDK 1.8.0_144, VM 25.144-b01
[info] # VM invoker: /Library/Java/JavaVirtualMachines/jdk1.8.0_144.jdk/Contents/Home/jre/bin/java
[info] # VM options: <none>
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: org.embulk.parser.xpath2.ParseBenchmark.run
[info] # Run progress: 0.00% complete, ETA 00:00:06
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: 0.062 ops/s
[info] # Warmup Iteration   2: 0.096 ops/s
[info] # Warmup Iteration   3: 0.095 ops/s
[info] Iteration   1: 0.099 ops/s
[info] Iteration   2: 0.093 ops/s
[info] Iteration   3: 0.096 ops/s
[info] Result "org.embulk.parser.xpath2.ParseBenchmark.run":
[info]   0.096 ±(99.9%) 0.062 ops/s [Average]
[info]   (min, avg, max) = (0.093, 0.096, 0.099), stdev = 0.003
[info]   CI (99.9%): [0.034, 0.159] (assumes normal distribution)
[info] # Run complete. Total time: 00:01:08
[info] Benchmark            Mode  Cnt  Score   Error  Units
[info] ParseBenchmark.run  thrpt    3  0.096 ± 0.062  ops/s
```

after

```
[info] Running (fork) org.openjdk.jmh.Main -i 3 -wi 3 -f1 -t 1
[info] # JMH version: 1.19
[info] # VM version: JDK 1.8.0_144, VM 25.144-b01
[info] # VM invoker: /Library/Java/JavaVirtualMachines/jdk1.8.0_144.jdk/Contents/Home/jre/bin/java
[info] # VM options: <none>
[info] # Warmup: 3 iterations, 1 s each
[info] # Measurement: 3 iterations, 1 s each
[info] # Timeout: 10 min per iteration
[info] # Threads: 1 thread, will synchronize iterations
[info] # Benchmark mode: Throughput, ops/time
[info] # Benchmark: org.embulk.parser.xpath2.ParseBenchmark.run
[info] # Run progress: 0.00% complete, ETA 00:00:06
[info] # Fork: 1 of 1
[info] # Warmup Iteration   1: 0.073 ops/s
[info] # Warmup Iteration   2: 0.146 ops/s
[info] # Warmup Iteration   3: 0.098 ops/s
[info] Iteration   1: 0.103 ops/s
[info] Iteration   2: 0.087 ops/s
[info] Iteration   3: 0.146 ops/s
[info] Result "org.embulk.parser.xpath2.ParseBenchmark.run":
[info]   0.112 ±(99.9%) 0.557 ops/s [Average]
[info]   (min, avg, max) = (0.087, 0.112, 0.146), stdev = 0.031
[info]   CI (99.9%): [≈ 0, 0.669] (assumes normal distribution)
[info] # Run complete. Total time: 00:00:59
[info] Benchmark            Mode  Cnt  Score   Error  Units
[info] ParseBenchmark.run  thrpt    3  0.112 ± 0.557  ops/s
[success] Total time: 62 s, completed 2017/10/24 16:10:47
```